move mcg upgrade tests to manage directory
